packages/agent/requirements.txt
anthropic>=0.43.0
fastapi>=0.115.0
uvicorn>=0.34.0
supabase>=2.0.0
realtime>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
